pyproject.toml
[project]
name = "2026-tgam-bocspeeches"
version = "0.1.0"
description = "Scraping and analysis of Bank of Canada speeches for hawkishness scoring"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.66.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "nbformat>=5.10.4",
    "pandas>=3.0.1",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
    "trueskill>=0.4.5",
]
